Fix unaddressed @senior-dev pings - eliminate closures, standardize error handling, clean up APIs (#86)
* Initial plan

* Add PartitionIndex variant and PartitionedSparseArray wrappers, update EntityRegistry and BehaviorRegistry


* Update RuleRegistry, SequenceRegistry, and SequenceDriver to use partitioned arrays


* Update all entity selectors to use PartitionedSparseArray and add TensorSparse overloads


* Delete RefBehaviorRegistry and IBehavior, remove IBehavior constraints from all behaviors


* Remove MaxEntities constant and ToInt() - prepare for proper partition handling


* Update all registries to use partitioned arrays - work in progress


* Fix HierarchyRegistry to enforce partition boundaries - no cross-partition parent-child relationships


* Fix iteration over partitioned arrays, update command drivers, attempt TryMatch fixes - WIP


* Fix PartitionedSparseArray/RefArray using Visit pattern - down to 504 errors


* Fix FlexRegistry partials and reduce errors to 50 - almost done


* Final attempt at TryMatch fixes - 62 errors remaining (down from 1172)


* AGENTS.md update

* Fix all TryMatch issues and add static lambdas - down to 1 error


* Complete partitioned sparse array split - zero compilation errors!


* Fix EntityRegistryIntegrationTests for partitioned indices


* Fix critical allocation bugs and add Count property - code review issues resolved


* Fix all test failures - ALL 386 TESTS PASSING! 🎉


* Fix closures in PartitionedSparseArray - use TryMatch instead of Visit, fix SparseArray to use uint capacity, refactor JsonNodeExtensions with guard clauses, split EntityRegistry and BehaviorRegistry functions


* Fix all FlexRegistry files - restore from passing commit e897921 where they were already correct


* Fix critical closure allocations - use TryMatch and static lambdas in hot paths


* Fix SparseArray to use uint indices throughout - remove ushort truncation bug


* WIP: Fix remaining review comments - tests failing due to SparseArray uint conversion issues


* Document SparseArray uint conversion pattern in AGENTS.md


* Fix SparseArray, SparseReferenceArray, SparseRef, and TensorSparse - remove int casts, use proper array initialization


* Fix JsonNodeExtensions - use ushort for global, delete TryGetEntityIndex, track partition in RulesDriver


* Fix EntityRegistry - add static keywords to lambdas, delete GetActiveEntities, update tests to use Concat of both partitions


* Fix FlexRegistry - remove null check after TryGetValue, add EnsureDirtyNode helper to all partials


* Fix HierarchyRegistry - use TryMatch, add TryGetChildrenArray, use extension methods, remove redundant null checks


* Fix final 3 review comments - ParentBehavior partition picking, RuleRegistry public field, SequenceRegistry TryActivate pattern


* Fix compiler warning in PropertyValueConverter - add null-forgiving operator for TryMatch string


* Add comments for breaking change tracking and enhancements

Added comments for future development regarding breaking changes and function enhancements.

* Update RuleRegistry.cs

* Add comments for .IsGlobal property implementation

Add comments for future enhancements regarding .IsGlobal property and extension method.

* Fix migration logic in FlexRegistry.FlexZOverride

Updated migration logic to use EnsureDirtyNode method.

* Fix migration logic in FlexRegistry.AlignContent

Updated migration logic to use EnsureDirtyNode method.

* Address all @senior-dev pings - add IsGlobal properties, TryGetEntityIndex, TryActivate pattern, clean up RulesDriver partition tracking


* Fix critical allocation issue - replace List<bool> with pre-allocated bool[] array in RulesDriver, cache IsGlobal in ParentBehavior, document TryActivate behavior in SceneLoader


* Update FlexRegistry.AlignContent.cs

* Add migration comments in FlexRegistry.FlexZOverride.cs

Add migration reminder comments for FlexZOverride file.

* Update EntityExtensions.cs

* Refactor child tracking logic and add comments

Refactor child tracking logic to use TryMatch method for improved clarity and consistency. Added comments for senior developer review.

* Update HierarchyRegistry.cs

* Address all @senior-dev pings - use TryMatch, ErrorEvents, IsGlobal, delete GetChildrenArray, use EnsureDirtyNode, update ParentBehavior.TryFindParent signature


* Fix benchmark build errors - cast uint indices to int


* CRITICAL FIX: Remove dangerous uint to ushort cast in scene entity activation - MaxSceneEntities is configurable!


---------

+++ MonoGame/Atomic.Net.MonoGame.Benchmarks/Core/SparseArrayBenchmark.cs
@@ -19,14 +19,14 @@ public class SparseArrayBenchmark
     public void Setup()
     {
         _array = new int?[Capacity];
-        _sparseArray = new SparseArray<int>((ushort)Capacity);
+        _sparseArray = new SparseArray<int>((uint)Capacity);
 
-        for (ushort i = 0; i < Capacity; i++)
+        for (uint i = 0; i < Capacity; i++)
         {
             if (_rng.NextDouble() < FillPercent)
             {
-                _array[i] = i;
-                _sparseArray.Set(i, i);
+                _array[i] = (int)i;
+                _sparseArray.Set(i, (int)i);
             }
         }
     }
@@ -57,7 +57,7 @@ public class SparseArrayBenchmark
         foreach(var (i, value) in _sparseArray)
         {
             result += value;
-            result += i;
+            result += (int)i;
         }
 
         return result;
